Rename sync-focused rust source file
